[android] fix fullscreen exit on Route From
Fixes #8915

When on fullscreen mode, after selecting "Route to", the app leaves
fullscreen mode when initiating the navigation window. The same was
not happening when navigating in "Route from" mode, and it was only
possible to leave fullscreen mode after leaving the navigation window.

This commit fixes that bug, ensuring that the app leaves fullscreen
mode when entering "Route from" navigation. Fullscreen mode can then
be set again after exiting navigation.

Signed-off-by: David Toma <david.toma@tecnico.ulisboa.pt>