[android] Fix FeatureId check in MapObject
This commit fix the way `FeatureId` is checked in `MapObject.equals()`.

Instead of checking if `mFeatureId` and `other.getFeatureId()` are not `FeatureId.EMPTY`, it now uses a new `isRealId()` method in `FeatureId` to determine if the IDs are valid for comparison.

The `isRealId()` method checks if the `mMwmName` is not empty, `mMwmVersion` is greater than 0, and `mFeatureIndex` is greater than 0. This provides a more robust check for valid feature IDs.

Signed-off-by: Mihail Mitrofanov <mitrofanov@bitrix.ru>